Create README.md
## Commit Description

In this commit, I've added a comprehensive README.md file to provide an informative and organized introduction to my GitHub repository. The README is designed to give users a clear understanding of the projects and their contents, facilitating easier navigation and utilization of the machine learning prediction projects stored in this repository.

### Key Changes Made:

- Added a descriptive title for the repository, "Machine Learning Prediction Projects Repository," providing clarity about the repository's purpose.
- Introduced myself with my name, Hiran Joseph, and provided my contact email for any inquiries or collaboration opportunities.
- Gave an overview of the repository's projects and their importance.
- Highlighted the machine learning projects included, specifically focusing on Linear Regression, Multiple Regression, and Outlier Treatment in Data Analysis.
- Included a "Getting Started" section to guide users on how to clone and access the repository's contents, ensuring that they can easily explore and work with the projects.
- Mentioned that the datasets used are included in each project folder or linked to external data sources.
- Clarified the licensing information to indicate that the repository and its projects are under the MIT License.
- Encouraged contributions and ideas from other users, signaling an open and collaborative environment for further development and enhancement of the projects.

This README.md serves as a central point of reference for anyone interested in my machine learning projects, making it easier for users to understand the purpose of this repository and how to get started with the various projects. It aims to enhance the accessibility and utility of the content and welcomes contributions and feedback from the GitHub community.